Refactor description code to a separate module.
